+++ spec/Pim/Bundle/CatalogBundle/EventSubscriber/MongoDBODM/EnsureIndexesSubscriberSpec.php
@@ -7,7 +7,7 @@ use PhpSpec\ObjectBehavior;
 use Pim\Bundle\CatalogBundle\Doctrine\MongoDBODM\IndexCreator;
 use Pim\Bundle\CatalogBundle\Doctrine\MongoDBODM\IndexPurger;
 use Pim\Bundle\CatalogBundle\Model\ChannelInterface;
-use Pim\Bundle\CatalogBundle\Entity\Currency;
+use Pim\Bundle\CatalogBundle\Model\CurrencyInterface;
 use Pim\Bundle\CatalogBundle\Model\LocaleInterface;
 use Pim\Bundle\CatalogBundle\Model\AttributeInterface;
 
@@ -226,7 +226,7 @@ class EnsureIndexesSubscriberSpec extends ObjectBehavior
 
     function it_generates_indexes_for_currency_when_enablig_it(
         $indexCreator,
-        Currency $usd,
+        CurrencyInterface $usd,
         LifecycleEventArgs $args
     ) {
         $usd->isActivated()->willReturn(true);
@@ -239,7 +239,7 @@ class EnsureIndexesSubscriberSpec extends ObjectBehavior
 
     function it_removes_indexes_for_currency_when_disabling_it(
         $indexPurger,
-        Currency $usd,
+        CurrencyInterface $usd,
         LifecycleEventArgs $args
     ) {
         $usd->isActivated()->willReturn(false);