+++ src/Pim/Bundle/TransformBundle/Normalizer/MongoDB/ProductMediaNormalizer.php
@@ -3,7 +3,7 @@
 namespace Pim\Bundle\TransformBundle\Normalizer\MongoDB;
 
 use Akeneo\Bundle\StorageUtilsBundle\MongoDB\MongoObjectsFactory;
-use Pim\Bundle\CatalogBundle\Model\ProductMedia;
+use Pim\Bundle\CatalogBundle\Model\ProductMediaInterface;
 use Symfony\Component\Serializer\Normalizer\NormalizerInterface;
 
 /**
@@ -31,7 +31,7 @@ class ProductMediaNormalizer implements NormalizerInterface
      */
     public function supportsNormalization($data, $format = null)
     {
-        return ($data instanceof ProductMedia && ProductNormalizer::FORMAT === $format);
+        return ($data instanceof ProductMediaInterface && ProductNormalizer::FORMAT === $format);
     }
 
     /**