+++ src/Pim/Bundle/TransformBundle/Normalizer/Structured/AttributeGroupNormalizer.php
@@ -2,7 +2,6 @@
 
 namespace Pim\Bundle\TransformBundle\Normalizer\Structured;
 
-use Pim\Bundle\CatalogBundle\Entity\AttributeGroup;
 use Pim\Bundle\CatalogBundle\Model\AttributeGroupInterface;
 use Symfony\Component\Serializer\Normalizer\NormalizerInterface;